// apps/sync-service/src/openapi/order-dropping.ts
export type PostV2OrderDroppingBodyOrderComment = {
  /** @nullable */
  comment?: string | null;
} | null;

export type PostV2OrderDroppingBodyOrderItemsItemBatchesItem = {
  activity_id: number;
  code: string;
  expired_date: string;
  production_date: string;
  manufacture_name: string;
  qty: number;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  source_material_id?: number | null;
  /** @nullable */
  total_price?: number | null;
  /** @nullable */
  year?: string | null;
};

export type PostV2OrderDroppingBodyOrderItemsItemStocksItem = {
  id?: number;
  qty?: number;
  activity_id?: number;
};

export type PostV2OrderDroppingBodyOrderItemsItem = {
  material_code: string;
  batches?: PostV2OrderDroppingBodyOrderItemsItemBatchesItem[];
  stocks?: PostV2OrderDroppingBodyOrderItemsItemStocksItem[];
};

export type PostV2OrderDroppingBody = {
  type: number;
  /** @nullable */
  required_date?: string | null;
  activity_id: number;
  /** @nullable */
  order_comment?: PostV2OrderDroppingBodyOrderComment;
  order_items: PostV2OrderDroppingBodyOrderItemsItem[];
  is_allocated: number;
  /** @nullable */
  is_manual?: number | null;
  vendor_code: string;
  customer_code: string;
  /** @nullable */
  delivery_number?: string | null;
  /** @nullable */
  no_po?: string | null;
  /** @nullable */
  service_type?: number | null;
};